move scan phase to seperate file
